+++ geofront/backends/github.py
@@ -104,6 +104,7 @@ class GitHubOrganization(Team):
     ACCESS_TOKEN_URL = 'https://github.com/login/oauth/access_token'
     USER_URL = 'https://api.github.com/user'
     ORGS_LIST_URL = 'https://api.github.com/user/orgs'
+    TEAMS_LIST_URL = 'https://api.github.com/user/teams'
 
     @typed
     def __init__(self, client_id: str, client_secret: str, org_login: str):
@@ -183,9 +184,24 @@ class GitHubOrganization(Team):
             response = request(identity, self.ORGS_LIST_URL)
         except IOError:
             return False
-        if isinstance(response, collections.Mapping) and 'error' in response:
+        if isinstance(response, collections.abc.Mapping) and \
+           'error' in response:
             return False
         return any(o['login'] == self.org_login for o in response)
+
+    def list_groups(self, identity: Identity):
+        if not issubclass(identity.team_type, type(self)):
+            return frozenset()
+        try:
+            response = request(identity, self.TEAMS_LIST_URL)
+        except IOError:
+            return frozenset()
+        if isinstance(response, collections.abc.Mapping) and \
+           'error' in response:
+            return frozenset()
+        return frozenset(t['id']
+                         for t in response
+                         if t['organization']['login'] == self.org_login)
 
 
 class GitHubKeyStore(KeyStore):